fix(vara.eth/ethexe): share crypto ops with gtest, fix Vara bench match
Move native gr_crypto implementations into ethexe-runtime-common
(std-gated crypto_ops) so both ethexe-processor and the gtest ethexe
runtime use one implementation; cover SyscallName::Crypto in the
pallet-gear syscall integrity bench match (#5456).
